Backend/requirements.txt
aiofiles==25.1.0
aiosmtplib==2.0.2
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.2
fastapi-mail==1.4.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
openai==1.12.0
pandas>=1.5.0
passlib[bcrypt]==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic-core==2.41.5
pypdfium2==5.0.0
python-docx==1.1.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.0
sqlalchemy==2.0.44
sqlmodel==0.0.16
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing-extensions==4.15.0
uvicorn[standard]==0.38.0
watchfiles==1.1.1
websockets==15.0.1
WTForms==3.1.2
